+++ src/main.ts
@@ -8,16 +8,34 @@ import {
 } from '@pulumi/pulumi/automation';
 import invariant from 'ts-invariant';
 import YAML from 'yaml';
-import { Commands, makeConfig } from './config';
+import { Commands, Config, InstallationConfig, makeConfig, makeInstallationConfig } from './config';
 import { environmentVariables } from './libs/envs';
 import { handlePullRequestMessage } from './libs/pr';
 import * as pulumiCli from './libs/pulumi-cli';
 import { login } from './login';
 
 const main = async () => {
+  const downloadConfig = makeInstallationConfig();
+  if (downloadConfig.success) {
+    await installOnly(downloadConfig.value);
+    core.info("Pulumi has been successfully installed. Exiting.");
+    return;
+  }
+
+  // If we get here, we're not in install-only mode.
+  // Attempt to parse the full configuration and run the action.
   const config = await makeConfig();
   core.debug('Configuration is loaded');
+  runAction(config);
+};
 
+// installOnly is the main entrypoint of the program when the user
+// intends to install the Pulumi CLI without running additional commands.
+const installOnly = async (config: InstallationConfig): Promise<void> => {
+  await pulumiCli.downloadCli(config.pulumiVersion);
+}
+
+const runAction = async (config: Config): Promise<void> => {
   await pulumiCli.downloadCli(config.options.pulumiVersion);
   await login(config.cloudUrl, environmentVariables.PULUMI_ACCESS_TOKEN);